Parse PV even after "bestmove" has been received
Fixes #9. 

Issue was the result of the bestmove being received before the final PV. This would cause the PV to no longer be parseable. This is resolved by making a PGN copy of the board, and undoing the last move if the PV is for the wrong side. 

Along with this, a single move number was being stored which caused PV updates after a move was made to report the wrong move number. These values are now tracked on a per-color basis to prevent this issue as well.