ACK messages immediately.
Once a message is received, ack it's ID immediately to prevent a
duplicate receive. On top of this, if IDs are received out of order log
this for future debugging.

Along with this, rather than queuing message processing indefinitely,
setup a timeout and let the queue build until processing can catch up.

Add port logging identifier for all messages for `grep` usage.

Fixes #52